Add ValueCollectionBase as base class for ValueArray and ValueRow
